refactor command handling
