+++ src/lib/permissions/permissions.ts
@@ -1,0 +1,3 @@
+export function runAllChecks(state: any) {
+    console.log(state)
+}
