+++ src/listeners/message.ts
@@ -1,10 +1,10 @@
 import type { ChatUserstate } from 'tmi.js'
-import { resolver as resolve } from '../lib/commands/commandResolver'
+import commandHandler from '../lib/commands/commandHandler'
 
 export function onMessage(channel: string, state: ChatUserstate, message: string, self: boolean) {
 	if (self) return
 	// TODO: command handler
 	if(message.trim().startsWith('!')) {
-		resolve(channel, state, message)
+		commandHandler.runCommand(channel, message)
 	}
 }
